requirements.txt
aiofiles==23.1.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
distro==1.8.0
Django==4.2.1
dlib==19.24.2
dnspython==2.3.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.2
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
googleapis-common-protos==1.59.0
graphviz==0.20.1
grpc-tools==0.0.1
grpcio==1.54.0
grpcio-tools==1.54.0
html5tagger==1.3.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
minio==7.1.14
multidict==6.0.4
numpy==1.24.3
opencv-python==4.7.0.72
packaging==21.3
paramiko==2.12.0
Pillow==9.5.0
protobuf==4.22.3
pycparser==2.21
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-consul==1.1.0
python-dotenv==0.21.0
PyYAML==5.4.1
redis==4.5.5
requests==2.28.1
sanic==23.3.0
sanic-routing==22.8.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==2.0.15
sqlparse==0.4.4
texttable==1.6.5
tracerite==1.1.0
typing_extensions==4.6.2
ujson==5.7.0
urllib3==1.26.12
uvloop==0.17.0
websocket-client==0.59.0
websockets==11.0.3
Werkzeug==2.3.4
zipp==3.15.0
